csv-data-lake/requirements.txt
pandas
pyarrow
python-dotenv
minio